+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
+- Added view submodule for building view objects
+- Added geometry functions (translation, rotation, reflection, dihedral angles, etc.)
 
 ## [0.0.7] - 2026-04-08
 - Added inertia moments, kabsch alignment, and center of mass algebraic methods to geom.py